Remove old code, put working code into unarchiver package
