--- src/main/java/com/gamebuster19901/excite/modding/Checked.java
@@ -1,7 +1,0 @@
-package com.gamebuster19901.excite.modding;
-
-public interface Checked {
-
-	public void check() throws AssertionError;
-	
-}
