+++ src/main/java/com/gamebuster19901/excite/modding/quicklz/QuickLZImpl.java
@@ -1,4 +1,4 @@
-package com.gamebuster19901.excite.modding.unarchiver;
+package com.gamebuster19901.excite.modding.quicklz;
 
 import java.nio.file.Path;
