+++ src/main/java/com/gamebuster19901/excite/modding/quicklz/package-info.java
@@ -1,0 +1,1 @@
+package com.gamebuster19901.excite.modding.quicklz;
