+++ src/main/java/com/gamebuster19901/excite/modding/unarchiver/Archive.java
@@ -1,136 +1,128 @@
 package com.gamebuster19901.excite.modding.unarchiver;
 
+import com.gamebuster19901.excite.modding.game.file.kaitai.ResMonster;
+import com.gamebuster19901.excite.modding.game.file.kaitai.TocMonster;
+
 import java.io.IOError;
 import java.io.IOException;
-import java.nio.ByteBuffer;
 import java.nio.file.Files;
 import java.nio.file.Path;
-import java.nio.file.Paths;
-import java.nio.file.StandardCopyOption;
-import java.nio.file.StandardOpenOption;
+import java.time.Instant;
+import java.util.Arrays;
+import java.util.LinkedHashMap;
+import java.util.List;
+import java.util.SequencedCollection;
 
-import com.gamebuster19901.excite.modding.FileUtils;
-import com.gamebuster19901.excite.modding.util.ByteBufUtils;
-import com.thegamecommunity.excite.modding.game.file.ResourceDecompressor;
-import com.thegamecommunity.excite.modding.game.file.ResourceFile;
+public class Archive {
 
-public class Archive implements ResourceFile {
+	private final Path archiveFile;
+	private final ResMonster archive;
+	private final Toc toc;
+	private byte[] bytes;
+	private final LinkedHashMap<String, ArchivedFile> files = new LinkedHashMap<>();
 	
-	private static final Path CURRENT_DIR = Paths.get("").resolve("run").toAbsolutePath();
+	public Archive(Path archivePath, Path tocPath) throws IOException {
+		this(archivePath, new Toc(tocPath));
+	}
 	
-	public static final Path DECOMPRESS_LOC = CURRENT_DIR.resolve("decomp");
-	public static final Path OUTPUT = CURRENT_DIR.resolve("out");
-	
-	static {
-		try {
-			FileUtils.deleteRecursively(DECOMPRESS_LOC);
-		} catch (IOException e) {
-			e.printStackTrace();
+	public Archive(Path archivePath, Toc toc) throws IOException {
+		this.archiveFile = archivePath;
+		this.archive = ResMonster.fromFile(archivePath.toAbsolutePath().toString());
+		this.toc = toc;
+		for(TocMonster.Details fileDetails : getFileDetails()) {
+			try {
+				files.put(fileDetails.name(), new ArchivedFile(fileDetails, this));
+			}
+			catch(Throwable t) {
+				//swallo
+			}
 		}
-		try {
-			FileUtils.deleteRecursively(OUTPUT);
-		} catch (IOException e) {
-			e.printStackTrace();
+	}
+	
+	public Toc getToc() {
+		return toc;
+	}
+	
+	public Instant getCreationDate() {
+		return toc.getCreationDate();
+	}
+	
+	public long getFileCount() {
+		return toc.getFileCount();
+	}
+	
+	public long getCompressedSize() {
+		return toc.getCompressedSize();
+	}
+	
+	public long getUncompressedSize() {
+		return toc.getUncompressedSize();
+	}
+	
+	public SequencedCollection<ArchivedFile> getFiles() {
+		return files.sequencedValues();
+	}
+	
+	public ArchivedFile getFile(String name) {
+		return files.get(name);
+	}
+	
+	public List<TocMonster.Details> getFileDetails() {
+		return toc.getFiles();
+	}
+	
+	public List<TocMonster.Filenames> getFileNames() {
+		return toc.getFileNames();
+	}
+	
+	public long getHash() {
+		return archive.header().hash();
+	}
+	
+	public boolean isCompressed() {
+		if (archive.header().compressed() == 128) {
+			return true;
 		}
-		
-		try {
-			Files.createDirectories(DECOMPRESS_LOC);
-			Files.createDirectories(OUTPUT);
-		} catch (IOException e) {
-			IOError e2 = new IOError(e);
-			throw e2;
+		else if (archive.header().compressed() == 0) {
+			return false;
 		}
-		
-	}
-
-	private final String name;
-	private final Path file;
-	
-	protected Archive(Path file) {
-		this.file = file;
-		this.name = file.getFileName().toString();
+		else {
+			throw new IOError(new IOException("Unknown compression value in header: " + archive.header().compressed()));
+		}
 	}
 	
-	@Override
-	public ByteBuffer getRawBytes() throws IOException {
-		return ByteBuffer.wrap(Files.readAllBytes(file));
-	}
-
-	@Override
-	public ByteBuffer getResourceBytes() throws IOException {
-		return this.getRawBytes().position(0x100).slice();
-	}
-
-	@Override
-	public ByteBuffer getHeaderBytes() throws IOException {
-		return this.getRawBytes().slice(0, 0x40);
-	}
-
-	@Override
-	public boolean isCompressedArchive() throws IOException {
-		int compress = this.getHeaderBytes().position(44).getInt();
-		System.out.println("compress: " + compress) ;
-		return compress != 0;
-	}
-	
-	public final Path getStrippedDest() throws IOException {
-		Path dest = Archive.DECOMPRESS_LOC.resolve(this.getName());
-		Path strippedDest = dest.resolve(this.getName() + ".stripped");
-		return strippedDest;
-	}
-	
-	public final Path getDecompDest() throws IOException {
-		Path dest = Archive.DECOMPRESS_LOC.resolve(this.getName());
-		Path decompDest = dest.resolve(this.getName() + ".decomp");
-		return decompDest;
-	}
-	
-	public void strip() throws IOException {
-		Files.createDirectories(getDecompDest().getParent());
-		Files.write(getStrippedDest(), ByteBufUtils.getRemaining(getResourceBytes()), StandardOpenOption.CREATE);
-	}
-	
-	public DecompressedArchive toDecompressedArchive() throws IOException {
-		strip();
-		Files.copy(getStrippedDest(), getDecompDest(), StandardCopyOption.REPLACE_EXISTING);
-		return DecompressedArchive.of(this);
-	}
-	
-	public static Archive of(Path archive) throws IOException {
-		Archive ret;
-		if(Files.isRegularFile(archive)) {
-			ret = new Archive(archive);
-			if(ret.isCompressedArchive()) {
-				System.out.println(archive.getFileName() + " is compressed!");
-				ret = new CompressedArchive(ret);
+	public byte[] getBytes() {
+		if(bytes == null) {
+			if(isCompressed()) {
+				bytes = archive.data().compressedData().bytes();
 			}
 			else {
-				System.out.println(archive.getFileName() + " is not compressed!");
+				bytes = archive.data().uncompressedData();
 			}
-			return ret;
 		}
-		throw new IllegalArgumentException(archive + " is not a file!");
+		return Arrays.copyOf(bytes, bytes.length);
 	}
 	
-	public Path getFile() {
-		return file;
+	public void validate() throws AssertionError {
+		if(archive.header().unixTimestamp() != toc.getCreationDate().getEpochSecond()) {
+			throw new AssertionError("Creation date mismatch!");
+		}
 	}
 	
-	public String getName() {
-		return name;
+	public Path getTocFile() {
+		return toc.getFile();
 	}
 	
-	private static String getFileName(Path f) {
-		 String fileName = f.getFileName().toString();
-		 int i = fileName.lastIndexOf('.');
-		 return (i == -1) ? fileName : fileName.substring(0, i);
+	public Path getArchiveFile() {
+		return archiveFile;
 	}
-
-	@Override
-	public ResourceDecompressor getDecompressor() {
-		return () -> {
-			return toDecompressedArchive().getDecompDest();
-		};
+	
+	public void writeTo(Path directory) throws IOException {
+		directory = directory.resolve(archiveFile.getFileName());
+		Path dir = Files.createDirectories(directory);
+		for(ArchivedFile file : files.sequencedValues()) {
+			file.writeTo(dir);
+		}
 	}
-
+	
 }
