--- src/main/java/com/gamebuster19901/excite/modding/util/TriPredicate.java
@@ -1,23 +1,0 @@
-package com.gamebuster19901.excite.modding.util;
-
-import java.util.Objects;
-
-@FunctionalInterface
-public interface TriPredicate<T, U, V> {
-
-	public boolean test(T t, U u, V v);
-	
-	default TriPredicate<T, U, V> and(TriPredicate<? super T, ? super U, ? super V> other) {
-		Objects.requireNonNull(other);
-		return (T t, U u, V v) -> test(t, u, v) && other.test(t, u, v);
-	}
-	
-	default TriPredicate<T, U, V> negate() {
-		return (T t, U u, V v) -> !test(t, u, v);
-	}
-	
-	default TriPredicate<T, U, V> or(TriPredicate<? super T, ? super U, ? super V> other) {
-		Objects.requireNonNull(other);
-		return (T t, U u, V v) -> test(t, u, v) || other.test(t, u, v);
-	}
-}
