--- src/main/java/com/gamebuster19901/excite/modding/util/package-info.java
@@ -1,1 +1,0 @@
-package com.gamebuster19901.excite.modding.util;
